DeathTest updates
Per googletest documentation[^1], it is strongly recommended that
death tests have names ending in `DeathTest` rather than simply
`Test`. This makes those changes.

Some death tests behave better when run in a threadsafe
manner[^2]. This is a trade-off that costs speed. The flag has been
set in tests where it made a difference.

[^1]: https://github.com/google/googletest/blob/main/docs/advanced.md#death-test-naming
[^2]: https://github.com/google/googletest/blob/main/docs/advanced.md#death-test-styles

Signed-off-by: Alejandro R. Sedeño <asedeno@google.com>