set is_ssl on upstream http filters with the upstreamInfo
The downstream headers are being passed to the upstream headers and if is_ssl is true - it adds the relevant header to the scheme.
This is_ssl flag is set according to the downstream info - meaning if the downstream is using ssl - so will the upstream.
The problem arises when we want to use http connection from the downstream and https on upstream or the other way around.
This fix will make it so we will take the is_ssl according to the
upstream info.

adding all usecases in upstream.cc to take the sslConnection info from:
upstreamInfo()->upstreamSslConnection()
instead of the downstreamAddressProvider()

Signed-off-by: hogery <doron.hogery@gmail.com>