Downgrade advisory log message
There is nothing inherently wrong with having `strip_query` set when having a path redirect that has its own query. The `strip_query` is a noop in such case, but its presence does not negatively affect anything, and there is nothing special this code does in such a case other than log the fact. 

In our massively multitenant Envoy deployment, users are allowed to specify both strip_query and redirect path with its own query, and many users do, which leads to a lot of log spam from this WARNING.

Logging this at DEBUG level is also consistent with other purely advisory logs in this class. 

Signed-off-by: Yury Kats <ykats@google.com>