Relax recent SNI restrictions (#36950)
See https://github.com/istio/istio/issues/53426. Istio has used
underscores in their SNI since the beginning and it is critical to its
functionality. Usage of underscores in SNI is a bit of a grey area in
the RFCs, which are extremely under-specified wrt to what exactly is the
allowed formats. However, the de-facto standard is to allow them, as
virtually every TLS library does so (including, but not limited to,
Golang, rustls, openssl, boringssl).

This PR loosens the restriction to additionally allow underscores.

Note the intent of the SNI restrictions was not RFC compliance, etc --
but rather to fix [log
injection](https://github.com/envoyproxy/envoy/security/advisories/GHSA-p222-xhp9-39rc)
attacks (putting ANSI escapes, HTML, etc) into logs. This change does
not loosen the security properties we hoped to gain with the initial
patch.

Signed-off-by: John Howard <john.howard@solo.io>