[postgres] Remove <> after constructor in the PG proxy
Angle brackets are not required after constructor and,
maybe, aren't even correct, though I'm not 100% sure on
what the standard says on the matter.

It seems like clang is fine with this syntax, but when you
try to build Envoy with gcc it complains:

```
./contrib/postgres_proxy/filters/network/source/postgres_message.h: At global scope:
./contrib/postgres_proxy/filters/network/source/postgres_message.h:397:14: error: expected unqualified-id before ')' token
  397 |   Sequence<>() = default;
      |              ^
Target //contrib/exe:envoy-static failed to build
```

Given that it's at least unusual to have angle brackets after
constructor in a class template specialization let's remove
them and satisfy both gcc and clang.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>