[contrib] Stop building ICU tools to fix gcc Envoy build
Clang and gcc are subtly different and it seems to be the cause
of contrib build failures reported in
https://github.com/envoyproxy/envoy/issues/31807 (e.g., when using
gcc to link the final binary it results in a bunch of essential
for gcc C++ symbols like __gxx_personality_v0).

This issue could be addressed (i.e., Envoy can be built using gcc).
However, given that it only happens when we build ICU binary tools
and those tools aren't used by Envoy directly or the build system,
we can just not build them and address the issue this way.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>