[contrib][http language filter] Change position of libstdc++ library when linking ICU tools
Clang and gcc are subtly different and it seems to be the cause
of contrib build failures reported in
https://github.com/envoyproxy/envoy/issues/31807 (e.g., when using
gcc to link the final binary it results in a bunch of essential
for gcc C++ symbols like __gxx_personality_v0).

The issue appear to be the order of the libraries when linking.
gcc, when building statically linked binaries basically needs
libstdc++ to be the last library or alsmot the last library in
the command line. And clang does not appear to care about it
much.

This change provides libstdc++ library in LIBS environment variabe
which will put it in the right position when building the ICU
library. This works well for both clang and gcc.

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>