rbac: add support for matching on route metadata (#36957)
## Background

This PR adds a new option called `sourced_metadata` to RBAC permissions
and principals which allows specifying an optional source for the
metadata to be matched. Currently it only supports Dynamic Metadata and
Route Metadata. More options could be added later.

**Fixes:** #34913

---

**Commit Message:** rbac: add support for matching on route metadata
**Additional Description:** This PR adds a new option called
`sourced_metadata` to RBAC permissions and principals which allows
specifying an optional source for the metadata to be matched.
**Risk Level:** Low
**Testing:** Added Unit & Integration Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>