Change handling of graceful case of LoadStatsReporting onRemoteClose (#37076)
Currently, we treat all remote grpc stream closes as errors, and log
warnings and increment failure metrics for every instance. A remote grpc
stream close with a 0 code is not a failure, but instead a graceful
termination, so it should be a lower log level and should not increment
failure metrics.

Risk Level: Low

---------

Signed-off-by: Brian Sonnenberg <bsonnenberg@google.com>

Signed-off-by: Brian Sonnenberg <111394771+briansonnenberg@users.noreply.github.com>