json: replacing IS_ENVOY_BUG when a large number value is used with an error (#37267)
Commit Message: json: replacing IS_ENVOY_BUG when a large number value
is used with an error
Additional Description:
Followup to #36919. In #36919, there was a behavior change when the JSON
library parsed a large number.
Prior to #36919 Envoy would have thrown an exception which ended up
rejecting the value or erroring out. After that an ENVOY_BUG was
introduced, but probably shouldn't have had, as Envoy may receive a
value that is large and should handle it correctly (e.g., when ingesting
a config).

This was detected due to fuzz bug
[379811166](https://g-issues.oss-fuzz.com/issues/379811166).

Risk Level: low
Testing: Added fuzz test case, and updated the unit-tests.
Docs Changes: N/A (no docs were updated in the original PR).
Release Notes: N/A (no release notes were introduced in the original
PR).
Platform Specific Features: N/A

Signed-off-by: Adi Suissa-Peleg <adip@google.com>