mobile/ci: Remove more unneeded iOS test apps (#37492)
One test app is sufficient to run on CI and validate that an iOS app can
be built and run on iOS platforms. The one app that will be used is
//test/swift/apps/experimental:app. It includes various types of test
filters for more comprehensive usage. The rest of the apps are deleted
to reduce the burden on CI's macos runners.

Signed-off-by: Ali Beyad <abeyad@google.com>