http: unified the filter chain creation in the FilterManager (#37112)
Commit Message: http: unified the filter chain creation in the
FilterManager
Additional Description:

This refactor ensure the all the http filter chain creation is done in
the FilterManager. (This also make sure we can do more control to the
filter chain creations.)

In the previous implementation, the `createFilterChain` method (of
`FilterManager`) is only used by the downstream filter chain. And the
upstream filter chain use another way to create filter chain. It make
that it's hard to control the behavior of filter chain creation in
uniform way.

The `createFilterChain` will return a value to tell whether the upgrade
is rejected. And use `upgradeFilterChainCreated` callback to tell
whether the upgrade is accepted. If both are none, then, that means no
upgrade.

In the new implementation, we unified the way of filter chain creation.
The `createFilterChain` method will be used for both downstream and
upstream. And a uniform way will be sued to report the upgrade results.

Risk Level: mid.
Testing: n/a.
Docs Changes: n/a.
Release Notes: n/a.
Platform Specific Features: n/a.

---------

Signed-off-by: wangbaiping(wbpcode) <wangbaiping@bytedance.com>