[redis_proxy] Add support for UNWATCH (#37620)
This PR adds support for the `UNWATCH` command in the Redis proxy.
Because of the lack of `UNWATCH` support, transactions are not well
supported by clients such as StackExchange.Redis, since conditions
require `UNWATCH` support.

**The Challenge**
The challenge with implementing `UNWATCH` in Envoy is knowing which
server to send the `UNWATCH` command to.
This was solved in this PR by considering `WATCH` a Transaction Command.
The behavior is explained below.

**Expected Behavior**

1. `WATCH` as a transaction command.
Making `WATCH` a transaction command means it will be handled by the
transaction handler, not the simple request handler. There, we set the
_transaction key_ in advance, without starting the transaction. This
means we will have state of the key where the latest `WATCH` was sent
to.

2. `UNWATCH` after `WATCH`
If an `UNWATCH` is sent right after a `WATCH`, we send it to the server
of the transaction key that was set. After that, we reset the key.
Covered by test: ✅

3. `UNWATCH` without key
Sending an `UNWATCH` without previous key being set, will have a direct
proxy response of `+OK\r\n`. Covered by test: ✅

4. `UNWATCH` after `MULTI` (no key)
Sending an `UNWATCH` without a key but within a transaction, will also
be a no-op in the proxy. But the response this time is `+QUEUED\r\n`, to
match Redis behavior. Covered by test: ✅

5. `UNWATCH` with key
Sending an `UNWATCH` with a transaction key set (either by `WATCH` or a
_simple command_ within an active transaction) will send the command to
the server of such hash key. Covered by test: ✅

6. `MULTI` after `WATCH`
Since `WATCH` now sets the transaction key in advance, in this scenario
we no longer queue the `MULTI` command, and we send it directly
upstream, based on the key set by the `WATCH` command. Covered by test:
✅

7. `MULTI` after `UNWATCH`
If an `UNWATCH` is sent after a `WATCH`, the transaction key is reset,
and the `MULTI` command is queued. This is the existing behavior in the
proxy. Covered by test: ✅

---------

Signed-off-by: Diogo Barbosa <diogobarbosa@microsoft.com>