Fix OTel GRPC trace exporter dropping spans (#37692)
Fixes otel exporter dropping spans in some cases, for example:
- If, while waiting for a response to an export request, the span buffer
is flushed again.
- If the span buffer is flushed after an upstream span ends, and there
exists a corresponding downstream span, and the buffer is flushed again
when the downstream span ends. This can be triggered by setting the
`tracing.opentelemetry.min_flush_spans` runtime variable to `1`, or if
the flush timer expires just after the upstream span ends. This is
similar to the first case, but the specific order of http callbacks
prevented the exporter client from ever being reset before the
downstream request completed, if it was used to export spans after the
upstream request completed.

Risk Level: 
Testing: Yes; added an integration test that I think will cover this.
Updated mock usage in other otel tests to account for refactors to the
exporter.

---------

Signed-off-by: Joe Kralicky <joekralicky@gmail.com>