mobile: Don't activate previous context on engine destruction
When the StrippedMainBase destructor is invoked in Envoy Mobile engine
destruction, occassionally the program crashes first due to trying to
activate the previous saved logging Context. There's a race condition
with accessing the static Context and its own destructor being invoked.

This change allows Envoy Mobile's use of StrippedMainBase to avoid those
crash issues by not trying to reactivate the previous saved logging
Context on engine destruction, since it wouldn't be useful to the library
anyway.

Signed-off-by: Ali Beyad <abeyad@google.com>