logger: Create only one Context per process for mobile
The Logger::Context operates as a stack. When a new one is created, it
becomes the active Context. When it gets destroyed, the destructor pops
the previous Context as the currently active one.

This can lead to lifetime bugs in Envoy Mobile, if multiple engines are
created and the destruction of those engines lead to indeterminate
destruction of the Logger::Context objects.

In this commit, the behavior is changed so that the Logger::Context
initialization is done by the external caller, not inside
StrippedMainBase as previously done. This allows the caller, in this
case, Envoy Mobile, to ensure only one Logger::Context is ever created,
circumventing any potential issues on Context destruction with popping
the previously saved context.

Signed-off-by: Ali Beyad <abeyad@google.com>