Drop unused constants
On newer version of Clang those generate a warning and with the settings
we use ultimately result in a compilation failure.

We probably should either disable the warning or just adjust the test.
Looking at the history of the test, it seems that these constants have
never been used, so I prefer to just drop them.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>