router: make shadow requests inherit parent sampling decision by default (#37874)
## Description

This PR changes the behavior of shadow request sampling so that if trace
sampling is not explicitly configured in the shadow policy, the shadow
request will inherit the parent's sampling decision.

After this change, sampling will follow the trace sampling policy of the
original request, which prevents oversampling when runtime sampling is
disabled.

**Fixes:** #37766

---

**Commit Message:** router: make shadow requests inherit parent sampling
decision by default
**Additional Description:** 
**Risk Level:** Low
**Testing:** Added Unit Tests
**Docs Changes:** Added
**Release Notes:** Added

---------

Signed-off-by: Rohit Agrawal <rohit.agrawal@databricks.com>