mobile: Fix order-dependent //test/common/logger:logger_delegate_test  (#37986)
Internally we discovered an order-dependent test that will cause the
test to fail because changing the log level is in `LogCbWithLevels` is
global. The fix is to restore the log level in the `SetUp`.

Risk Level: low (test only)
Testing: unit test
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Fredy Wijaya <fredyw@google.com>