Rfresh the version of hessian2-code dependency
The new version includes https://github.com/alibaba/hessian2-codec/pull/37
that contains a fix for one of the issues listeed in
https://github.com/envoyproxy/envoy/issues/37911 (number 5 on the list).

TL;DR: the issue is that in hessian2-codec the same cc file is used as
a source for two different libraries. The code that depends on both of
those libraries at the same time might have an issue because of the same
source used several times.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>