Fix contrib/cryptomb/private_key_providers/test:speed_test build in FIPS mode
Some of the tests in that file rely on EC_group_p256 function that was
first introduced in
https://github.com/google/boringssl/commit/417069f8b2fd6dd4f8c2f5f69de7c038a2397050.
Evnidently the version of BoringSSL that Envoy uses in FIPS mode does
not contain this change. Because of that build of speed_test.cc fails.

To fix the issue I'm replacing calls to EC_group_p256 with
EC_GROUP_new_by_curve_name(NID_X9_62_prime256v1). To the best of my
knowledge functionally it should be exactly the same thing.

NOTE: Probably a more long term solution would be to update the version
of BoringSSL we are using to newer one, as the current one is more than
2 years old at this point. However, it's a bit more work than I can do
right now, thuse this temporary fix. I will return to the question of
updating BoringSSL though affter I close the ongoing bugs.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>