http_inspector dynamic buffering (#37002)
Changed the buffering behavior of HttpInspector. Previously, a static
buffer of 8KB was inspected, and anything larger (e.g. a GET request
with >8KB of params) would fail. Now, there is a configurable
initial_read_buffer_size in HttpInspector that, and if the HTTP version
is not identified with that amount of request buffered, it will
iteratively double up to 64KB.

Additional Description:
Risk Level: low
Testing: TBD
Fixes issue #36162

---------

Signed-off-by: Brian Sonnenberg <bsonnenberg@google.com>

Signed-off-by: Brian Sonnenberg <111394771+briansonnenberg@users.noreply.github.com>