Fix broken refs and attention blocks
ReStructuredText has a way to cross-reference sources using the :ref: syntax.
It turns out that going over /api/ with regex `[^:]ref:` showed some hits, meaning
the syntax was not parsed and the link was broken.

The same for the text blocks like `.. attention::`, the text to be highlighted
needs to be indented with at least 2 spaces, otherwise the block is not interpreted,
this was done using `\/\/\s+\.\.\s\w+::\n\s+\/\/(\s{0,1}\w)`

Signed-off-by: Jens Kat <jenskat@gmail.com>