Fix ext_proc duplicate data issue (#37994)
Fixes: https://github.com/envoyproxy/envoy/issues/36119

The issue is due to ext_proc failed to send the last empty chunk with
end_stream = true to the backend server. This ends up causing a copy of
the data stored in the decoding buffer for the retry purpose are sent
again, which caused duplicate data is sent to backend server. The fix is
simply having the ext_proc filter inject the last empty chunk to the
filter chain and send to the backend server.

---------

Signed-off-by: Yanjun Xiang <yanjunxiang@google.com>