Refresh the version of tcmalloc
I'm trying to build Envoy with Clang-18 and the version of tcmalloc
we currently uses trips thread safety analysis, which generates a
warning and ultimately results in a build failure.

I could disable the warning, but I think thread safety analysis warnings
might be potentially useful. Plus, the issue is addressed in a later
version of tcmalloc and the version we use is over 2 years old by now.

I figured the best way to address the warning then might be to refresh
the version of tcmalloc Envoy uses.

This change bumps it to the commit that contained the fix for the
warning I want to address, though I'm not avert to using even more fresh
version either if there is interest in that.

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>