Patch tcmalloc to fix the thread safety analysis warning (#38120)
I'm trying to build Envoy with Clang-18 and the version of tcmalloc we
currently uses trips thread safety analysis, which generates a warning
and ultimately results in a build failure.

The issue was fixed in the upstream tcmalloc, however the current
version of binutils in the Envoy CI image not new enough to be able to
build fresh versions of tcmalloc.

Given that we are already in a process of updating LLVM toolchain
version in the Envoy CI, for now I'm going with patching tcmalloc
instead of addressing all the issues required to bump the version of
tcmalloc.

Additional Description: Related to the work in
https://github.com/envoyproxy/envoy/issues/37911