Docs: Fix broken refs and attention blocks (#38061)
ReStructuredText has a way to cross-reference sources using the :ref:
syntax. It turns out that going over `/api/` (excluding
`api/envoy/**/v2/*`) and `/docs` with regex `[^:]ref:` showed some hits,
meaning the syntax was not parsed and the link was broken.

The same for the text blocks like `.. attention::`, the text to be
highlighted needs to be indented with at least 2 spaces, otherwise the
block is not interpreted, this was done using
`\/\/\s+\.\.\s\w+::\n\s+\/\/(\s{0,1}\w)`
Risk Level: None
Docs Changes: 
-
https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/filters/network/tcp_proxy/v3/tcp_proxy.proto.html#extensions-filters-network-tcp-proxy-v3-tcpproxy
-
https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/filters/http/cors/v3/cors.proto.html#extensions-filters-http-cors-v3-cors
-
https://www.envoyproxy.io/docs/envoy/latest/api-v3/config/core/v3/health_check.proto.html#config-core-v3-healthcheck
-
https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/upstream/connection_pooling.html#happy-eyeballs-support
-
https://www.envoyproxy.io/docs/envoy/v1.33.0/configuration/http/http_filters/compressor_filter#statistics

---------

Signed-off-by: Jens Kat <jenskat@gmail.com>