Bump up version of the libipp-crypto dependency (#37940)
The currently used version of libipp-crypto library has an issue when we
build it using fresh versions of LLVM toolchain (specifically I tried
18). clang-14, that is currently used, is somewhat old-ish (the latest
official release is from mid 2022) and in #37911 I'm looking and making
changes that will make it possible to build Envoy with newer versions of
LLVM toolchain.

Now, when it comes to libipp-crypto library, the version Envoy currently
uses internally defines a `__INLINE` macro. Identifiers that start with
two underscores are typically reserved for compilers, standard library
and similar use cases, but it didn't cause problems with the older
version of LLVM toolchain. However, on LLVM-18 some system headers that
libipp-crypto relies on also define `__INLINE` macro (and then undefine
it) which causes conflicts and results in a compilation failure.

The issue has been addressed upstream and to fix it in Envoy we just
need to switch to a newer version of libipp-crypto to include
https://github.com/intel/cryptography-primitives/commit/ea7cd1566e8a24f214bd93e3318e6aa35a95b7e9
(and does not contain
https://github.com/intel/cryptography-primitives/commit/7d6ac349507258f49d00909df33d5dea4ff77f39).
This change bumps up the version of the libipp-crypto to a recent
release that contains the fix

In addition to newer clang versions I also did test that
`--define=boringssl=fips` builds work with existing clang version
(clang-14, `--config=docker-clang` and `--config=docker-clang-libc++`),
plus checked that the `--config=docker-gcc` still builds. I tested both
regular envoy and contrib versions, though this dependency is only used
by contrib. For the functional tests I rely on the tests that will be
run automatically by GitHub.

Related to #37911 

Fix #37628

Signed-off-by: Mikhail Krinkin <krinkin.m.u@gmail.com>
Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>